fix mypy issue: https://github.com/python/mypy/issues/7182
