+++ python/mallard-questionnaire-registry/mallard_qr/models.py
@@ -1,9 +1,7 @@
-from __future__ import unicode_literals
-
 from django.db import models
 from django.utils.translation import ugettext as _
 
-import aristotle_mdr as aristotle
+from aristotle_mdr import models as MDR
 
 
 """
@@ -11,22 +9,22 @@ These models are based on the DDI3.2 and the SQBL XML formats.
 """
 
 
-class AdministrationMode(aristotle.models.unmanagedObject):
+class AdministrationMode(MDR.unmanagedObject):
     pass
 
 
-class Question(aristotle.models.concept):
+class Question(MDR.concept):
     template = "mallard_qr/question.html"
     collected_data_element = models.ForeignKey(
-        aristotle.models.DataElement,
+        MDR.DataElement,
         blank=True, null=True, on_delete=models.SET_NULL,
         related_name="questions",
     )
-    question_text = aristotle.models.RichTextField(
+    question_text = MDR.RichTextField(
         blank=True,
         help_text=_("The text which describes the information which is to be obtained.")
     )
-    instruction_text = aristotle.models.RichTextField(blank=True)
+    instruction_text = MDR.RichTextField(blank=True)
     # administration_modes = models.ManyToManyField(AdministrationMode,blank=True,null=True)
     estimated_seconds_response_time = models.PositiveIntegerField(
         null=True, blank=True,
@@ -34,7 +32,7 @@ class Question(aristotle.models.concept):
     )
 
 
-class ResponseDomain(aristotle.models.aristotleComponent):
+class ResponseDomain(MDR.aristotleComponent):
 
     parent_field_name = 'question'
     class Meta:
@@ -42,7 +40,7 @@ class ResponseDomain(aristotle.models.aristotleComponent):
 
     question = models.ForeignKey(Question, related_name="response_domains", on_delete=models.CASCADE)
     value_domain = models.ForeignKey(
-        aristotle.models.ValueDomain,
+        MDR.ValueDomain,
         null=True, blank=True,
         on_delete=models.SET_NULL,
     )
@@ -64,17 +62,17 @@ class ResponseDomain(aristotle.models.aristotleComponent):
 
 
 """
-class QuestionModule(aristotle.models.concept):
+class QuestionModule(MDR.concept):
     template = "mallard-qr/questionmodule.html"
     questions = models.ManyToManyField(Question,blank=True,null=True)
     submodules = models.ManyToManyField('QuestionModule',blank=True,null=True)
-    instruction_text = aristotle.models.RichTextField(blank=True,null=True)
+    instruction_text = MDR.RichTextField(blank=True,null=True)
     sqbl_definition = TextField(blank=True,null=True)
     administration_modes = models.ManyToManyField(AdministrationMode,blank=True,null=True)
 
-class Questionnaire(aristotle.models.concept):
+class Questionnaire(MDR.concept):
     template = "mallard-qr/questionnaire.html"
     submodules = models.ManyToManyField(QuestionModule,blank=True,null=True)
-    instructionText = aristotle.models.RichTextField(blank=True)
+    instructionText = MDR.RichTextField(blank=True)
     administration_modes = models.ManyToManyField(AdministrationMode,blank=True,null=True)
 """
